--- packages/sdk/src/Portfolio/IntegrationManager.ts
@@ -1,13 +1,0 @@
-export {
-  Action,
-  Selector,
-  callEncode,
-  callDecode,
-  type CallArgs,
-  addTrackedAssetsEncode,
-  addTracketAssetsDecode,
-  type AddTracketAssetsArgs,
-  removeTrackedAssetsEncode,
-  removeTrackedAssetsDecode,
-  type RemoveTrackedAssetsArgs,
-} from "../_internal/IntegrationManager.js";
